==> 01_Summary_ECCC.rmd <==
---
title: "01_Summary_ECCC"
author: "Gen Perkins"
date: "2023-09-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

## Summary of ECCC

```{r library, warning=FALSE, message=FALSE, include = FALSE, echo = FALSE}
library(leaflet)
library(RColorBrewer)
library(dplyr)
library(lubridate)
library(sf)
library(adehabitatHR)
library(ggplot2)
library(stringr)
library(readxl)
library(dplyr)
library("rnaturalearth")
library("rnaturalearthdata")


#raw_dat <- file.path("data", "movebank_locations_20230407")
#raw_dat <- file.path("data", "movebank_locations_20230724")
#raw_dat <- file.path("data", "movebank_locations_20231006")
raw_dat <- file.path("data", "movebank_locations_20231219")

filesoi <- list.files(raw_dat)

# data_set3 : spring migration 
key = "ECCC"

filesoi <- list.files(raw_dat, pattern = key)
filesoi_ref <- filesoi[1]
filesoi <- filesoi[2]



# read in reference data 
brep <- read.csv(file.path(raw_dat, filesoi_ref))
brep <- brep %>%
      # dplyr::select(c(tag.id , "animal.id" ,deploy.on.date, "animal.marker.id","animal.ring.id", "tag.serial.no", animal.life.stage, deploy.on.measurements,  animal.mass,  tag.model, 
      #                 tag.manufacturer.name, tag.serial.no))%>%
mutate(tag.local.identifier = tag.id)
# 
# no.tags.ref <- unique(brep$tag.local.identifier)
# no.ids.ref <- unique(brep$animal.id)

# number of unique tags = no of ids 

# 121 tag ids 
# 71 names animal ids

# read in the location data 
btemp <- read.csv(file.path(raw_dat, filesoi))
bout <- btemp %>%
    dplyr::select(visible, timestamp, location.long, location.lat,
                  argos.altitude, import.marked.outlier, 
                  gps.fix.type.raw, lotek.crc.status,
                  sensor.type, individual.local.identifier,
                  argos.lc, algorithm.marked.outlier,
                  tag.local.identifier, height.above.ellipsoid)%>%
  mutate(date_time = ymd_hms(timestamp))

# 
# # calculate time differences
# bout <- bout  %>%
#   mutate(date_time = ymd_hms(timestamp))%>%
#   mutate(year = year(date_time )) %>%
#   mutate(month = month(date_time ),
#          day = day(date_time),
#          hour = hour(date_time),
#          minute = minute(date_time))
# 
# no.tags <- unique(bout$tag.local.identifier) # 121
# no.ids <- unique(bout$individual.local.identifier) # 62 + blanks 
#   

# 60 tracks and track names 
# 
# ids <- bout %>%
#   dplyr::select(tag.local.identifier, individual.local.identifier) %>%
#   group_by(tag.local.identifier, individual.local.identifier)%>%
#   summarise(n = n())
#   
# 
# ymin = min(range(bout$year))
# ymax = max(range(bout$year))


#output all data 

#brep #107
#bout

all_dat <- left_join(bout, brep )

head(all_dat)

all_dat <- all_dat %>% 
  dplyr::select(-individual.local.identifier, -tag.local.identifier, -animal.taxon, 
         -deploy.off.date, -import.marked.outlier, -algorithm.marked.outlier, 
         -height.above.ellipsoid) %>%
  mutate(tag.id = as.character(tag.id)) %>%
  filter() %>% 
  dplyr::mutate(lotek.crc.status = case_when(
    is.na(lotek.crc.status) ~ "",
    TRUE ~ as.character(lotek.crc.status)))%>%
  dplyr::filter(lotek.crc.status != "E")%>%
  dplyr::mutate(argos.lc = case_when(
    is.na(argos.lc) ~ "",
    TRUE ~ as.character(argos.lc)))%>%
  dplyr::filter(argos.lc != "Z")%>%
  filter(!is.na(location.long), 
         !is.na(location.lat))%>%
  mutate(tag.model = case_when(
    tag.model == "gps-pinpoint" ~ "Lotek PinPoint GPS-Argos 75", 
    tag.model == "PinPoint 75" ~ "Lotek PinPoint GPS-Argos 75", 
    tag.model ==  "Sunbird" ~ "Sunbird Solar Argos",
    tag.model ==  "sunbird" ~ "Sunbird Solar Argos",
    .default = as.character(tag.model)))



unique(all_dat$animal.id)

#"FORTESCU" "KIMBLESO" "MOORES"   "EASTPIT"  "PEIXE"    "NBRIG"   

all_dat <- all_dat %>% 
  dplyr::mutate(deploy.on.latitude = case_when(
    study.site == "FORTESCU" ~ 39.0322,
    study.site == "KIMBLESO" ~ 39.20075,
    study.site == "MOORES" ~ 39.20075,
     study.site == "EASTPIT" ~ 39.1958,
    study.site == "PEIXE" ~ -31.401,
    study.site == "NBRIG" ~ 39.436)) %>%
  dplyr::mutate(deploy.on.longitude = case_when(
        study.site == "FORTESCU" ~ -74.7948,
    study.site == "KIMBLESO" ~ -75.0255,
    study.site == "MOORES" ~ -75.0255,
    study.site == "EASTPIT" ~ -75.02589,
    study.site == "PEIXE" ~  -51.066,
    study.site == "NBRIG" ~ -74.346))
 
 
# # Thom = thompson beach 
# -75.0255 , 39.20075
#"#PEIXE"
#-31.401, -51.066




# #save out file
# #save out file
clean_save = all_dat %>% mutate(proj = "ECCC")
saveRDS(clean_save, file = file.path("output", "rekn_eccc_20240101.rds"))

```


This dataset has some discrepancies within the reference data and the track data. The reference data show 120 tag ids (based on tag-id number ie. 204561) + two individuals with no animal.ids. 
Of these I found deployment information on 71 animal individuals. The number of tracks was less. 

This dataset has `r length(no.tags)` tags (tag.local.identifier) and `r length(no.ids)` (individual.local.identifier). 

Note this may be a proportion of tags that are not yet deployed. See appendices. 

Note some records are not matched to the same tag.identifier( but this is only two few records)

```{r, echo = FALSE}
knitr::kable(ids)

```


```{r, echo=FALSE}

errors <- ggplot(bout, aes(year, fill = visible))+
    geom_bar(position = "dodge")+
    xlim(2021,2024)#+
  #facet_wrap(~tag.local.identifier)

clean <- bout %>%
  filter(visible == "true")%>%
  filter(!is.na(location.long)) %>%
  filter(!is.na(location.lat)) 

#length(clean$visible)

cleanargos <- clean %>% 
  dplyr::filter(argos.lc != "Z")%>%
  dplyr::filter(argos.lc != "A")%>% 
 dplyr::filter(argos.lc != "B")

#length(cleanargos$visible)

clean = cleanargos

errrors_tble <- bout %>%
  group_by(tag.local.identifier, year, visible) %>%
  summarise(count = n())

# #save out file
clean_save = clean %>% mutate(proj = "ECCC")
saveRDS(clean_save, file = file.path("output", "rekn_eccc_20231102.rds"))


```


# Raw data

The number of raw records as of September 2023 is `r length(bout$tag.local.identifier)`.  The raw date range varies from 
`r min(range(bout$year))` to `r max(range(bout$year))`.


```{r, echo = FALSE}
errors
```

If we remove all the algorithm marked outliers (lotek = error, argos accuracy value = A, B, Z). We reduce the number of observations in the data set from `r length(bout$tag.local.identifier)` to `r length(clean$tag.local.identifier)`.
Argos accuracy codes are: 

- Class 0: over 1500 m radius (not currently filtered)

- Class A: Argos no accuracy estimation (3 messages received)
- Class B: Argos no accuracy estimation (1 or 2 messages received)
- Class Z: Argos Invalid Location


The cleaned date range varies from 
`r min(range(clean$year))` to `r max(range(clean$year))`.


# GPS vs argos

The data split between gps and argos data records. Large proportion are argos doppler shift.   These represent two different methods of calculating the locations and all types can be used. 

```{r, echo = FALSE, message=FALSE}
tag_type <- clean %>%
  dplyr::select(tag.local.identifier, sensor.type) %>%
  group_by(tag.local.identifier, sensor.type)%>%
  summarise(n = n())

#tag_type

tag_type_date <- clean %>%
  dplyr::select(tag.local.identifier, sensor.type, year, month) %>%
   group_by(tag.local.identifier, sensor.type, year,month)%>%
  summarise(n = n())

#tag_type_date

p_alldat <- ggplot(clean, aes(year, fill = sensor.type))+
  geom_bar(position = "dodge") 

```

```{r, echo= FALSE}
p_alldat 

# p_alldat <- ggplot(clean, aes(tag.local.identifier))+ #, fill = sensor.type))+
#   geom_bar(position = "dodge") +
#   ylim(0, 100)+
#   #facet_wrap(~year)+
#   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

#library(scales)
#p_alldat + scale_y_continuous(trans = log2_trans())


```


# Duration of Tag summary

```{r, echo = FALSE}
# 
# p_duration <- ggplot(clean, aes(factor(month), fill = factor(year)))+
#   geom_bar(position = "dodge") +
#   #xlim(1,12)+
#   #facet_wrap(~tag.local.identifier)+
#   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
# 
# p_duration

table_max <- clean %>% 
  dplyr::select(tag.local.identifier,individual.local.identifier, date_time) %>%
 # group_by(tag.local.identifier)%>% 
  slice_max(date_time, by = tag.local.identifier)
colnames(table_max)<- c("tag.local.identifier", "individual.local.identifier","max")

table_min <- clean %>% 
#   dplyr::filter(year != 2025) %>%
  dplyr::select(tag.local.identifier,individual.local.identifier, date_time) %>%
 # group_by(tag.local.identifier)%>% 
  slice_min(date_time, by = tag.local.identifier)

colnames(table_min)<- c("tag.local.identifier", "individual.local.identifier", "min")

dur <- left_join(table_max, table_min, by = join_by(tag.local.identifier)) %>%
  distinct() %>%
  dplyr::mutate(duration = max- min)%>%
  mutate(dur_hrs = round(as.numeric(duration),1))%>%
  mutate(dur_days = round( dur_hrs/24,1))%>%
  mutate(year = year(min))

dur_data = dur
#library(ggplot2)

dur_plot <- ggplot(dur, aes(y=factor(tag.local.identifier))) +
  geom_segment(aes(x=min, xend=max, y=factor(tag.local.identifier), yend=factor(tag.local.identifier)), size=1)+
  xlab("Date") + ylab("Tag") 

dur_plot

dur_hist <- ggplot(dur, aes(x= dur_days, colour = as.factor(year)))+
  geom_histogram() + #fill="white", position="dodge") +
  facet_wrap(~year)+
  xlab("duration (days)") 

dur_hist

p_duration <- ggplot(clean, aes(factor(month), fill = factor(year)))+
  geom_bar(position = "dodge") +
  #xlim(1,12)+
  facet_wrap(~tag.local.identifier)+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))


```


Two individuals (232986 and 232985), have very long records on their tags. These are both sunbird tags deployed at Kimbles beach  and showed records in 2022 and 2023 in the same locations as previous years. Are these some feature of the sunbird tags? 


```{r,echo = FALSE}

e1 <- clean %>%
  filter(tag.local.identifier ==232986)

e1 <- e1 %>%
  dplyr::group_by(year , month)%>%
  dplyr::summarise( d_n = n())


e2  <- clean %>%
  filter(tag.local.identifier ==232985)

e2 <- e2 %>%
  dplyr::group_by(year , month)%>%
  dplyr::summarise( d_n = n())
  

```

```{r}
#individual = 232986
knitr::kable(e1)

# individual = 232985
knitr::kable(e2)
```

# Banding locations

The banding locations were largely in Delaware Bay (included KIMBLESO, MOORES, FORTESCUE,  EASTPIT? East point?) and a location in New Jersey (Pesca) along with a location in Brazil near Porto Alegra (PEIXE).
This table shows the number of individuals by banding location and year.

```{r, echo = FALSE}
#head(dur)

loc <- dur %>%
  dplyr::select(c(individual.local.identifier.x , dur_days))%>%
  mutate(dloc = stringr::str_sub(individual.local.identifier.x, 1, 4)) %>%
  mutate(dyr = stringr::str_sub(individual.local.identifier.x,13, 17)) %>%
  mutate(dyr = gsub("_", "", dyr))

dep_summary <- loc %>%
  dplyr::group_by(dloc, dyr)%>%
  dplyr::summarise( dloc_n = n())

#dep_summary

```

```{r, echo = FALSE}
knitr::kable(dep_summary)
```


# Types of tags

Of the tags deployed there was sunbird tags and argos pinpoint tags. THe sunbird tags lasted much longer approx 208 days on average (including the potential outliers). The argos tags were shorter in duration and generally captured the period of may/june to October. 

```{r, echo = FALSE}

loc <- dur_data %>%
  dplyr::select(c(individual.local.identifier.x , dur_days))%>%
  mutate(dloc = stringr::str_sub(individual.local.identifier.x, 1, 4)) %>%
  mutate(dyr = stringr::str_sub(individual.local.identifier.x,13, 17)) %>%
  mutate(dyr = gsub("_", "", dyr))

loc <- dplyr::left_join(loc, brep, by =c("individual.local.identifier.x" = "animal.id"))

loc <- loc %>% 
  dplyr::select(c(-deploy.on.date, -animal.ring.id, -tag.local.identifier, -tag.id))%>%
  mutate(tag.model = case_when(
    tag.model == "gps-pinpoint" ~ "Lotek PinPoint GPS-Argos 75",
    tag.model == "P08011 Sunbird Tag."~"sunbird",
    .default = as.character(tag.model)))


type_summary <- loc %>%
  dplyr::group_by(dloc,  tag.model)%>%
  dplyr::summarise( ids_n = n())


# type by duration
type_dur <- dur_data %>%
  dplyr::select(c(individual.local.identifier.x , dur_days))%>%
  mutate(dloc = stringr::str_sub(individual.local.identifier.x, 1, 4)) %>%
  mutate(dyr = stringr::str_sub(individual.local.identifier.x,13, 17)) %>%
  mutate(dyr = gsub("_", "", dyr))

type_dur <- dplyr::left_join(type_dur, brep, by =c("individual.local.identifier.x" = "animal.id"))%>%
  mutate(tag.model = case_when(
    tag.model == "gps-pinpoint" ~ "Lotek PinPoint GPS-Argos 75",
    tag.model == "P08011 Sunbird Tag."~"sunbird",
    .default = as.character(tag.model))) %>%
  dplyr::select(dur_days, tag.model, tag.local.identifier)


type_dur_sum <- type_dur %>%
  dplyr::group_by( tag.model)%>%
  dplyr::summarise( min_dur = min(dur_days), 
                    mean_dur = mean(dur_days),
                    max_dur = max(dur_days))

  

```

```{r, echo = FALSE}
knitr::kable(type_summary)
knitr::kable(type_dur_sum)
```

# Number of active tags 

At the last data download, August 1st, a number of tags were continuing to record locations. THese represent active tags where more details can be gained. Inactive tags are those which have stopped transmitting. 

```{r , echo=FALSE}
#head(dur_data)

library(dplyr)
ddur <- dur_data %>%
  dplyr::mutate(max_month = month(max),
         max_year = year(max))
ddur <- ddur %>%
  dplyr::mutate(active = ifelse(max_month == 9 & max_year == 2023, "y", "n"))

active = ddur %>% filter(active == "y")
inactive = ddur %>% filter(active == "n")

```


Of the `r length(ddur$tag.local.identifier)` tags, `r length(active$tag.local.identifier)` were active, and `r length(inactive$tag.local.identifier)` were inactive.

```{r, echo = FALSE}

#sort(inactive$dur_days)

ggplot(ddur, aes(x = dur_days, y = as.factor(tag.local.identifier), colour=  active))+ 
  geom_point()


sort(ddur$v)


ddur <- ddur %>%
  mutate(dur_class = case_when(
    dur_days < 29 ~ 1,
    dur_days >29 & dur_days <60 ~2, 
    dur_days >60.1 & dur_days <100 ~3, 
    dur_days >100 ~4,
    .default = NA
 ))

```


```{r, echo = FALSE}
#group the points into classes to help distinguish which can be used in the breeding area. 

#ggplot(ddur, aes(x = dur_days, y = as.factor(tag.local.identifier), colour=  as.factor(dur_class)))+ 
#  geom_point()
  
```

# Great Circle distance

Calculate the great circle distances to determine which individuals can be used for analysis

```{r}
#remotes::install_github("mike-decarlo/gcd")
library(gcd)

table_last <- clean %>% 
  #dplyr::filter(year != 2025) %>%
  slice_max(date_time, by = tag.local.identifier) %>%
  slice_head(n = 1, by = tag.local.identifier) %>%
  dplyr::select( location.long, location.lat, individual.local.identifier)#%>%
 # mutate(timept = "end")

table_first <- clean %>% 
  #dplyr::filter(year != 2025) %>%
  slice_min(date_time, by = tag.local.identifier) %>%
  dplyr::select( location.long, location.lat, individual.local.identifier)#%>%
 # mutate(timept = "start")

loc_table = left_join(table_first, table_last, by = 'individual.local.identifier')
# 
# hav_km <- dist_haversine(lon1 = loc_table[1,1], lat1 = loc_table[1,2]
#   , lon2 = loc_table[1,4], lat2 = loc_table[1,5], km = TRUE)
# 
# hav_km
# 

library(tidyverse)
#df <- data.frame(a = 1:10, b = 11:20, c = 21:30)

hav_km <- loc_table %>% 
    group_nest(row_number()) %>% 
    pull(data) %>% 
    map(function(x) transmute(x,
                                 var1 = dist_haversine(lon1 = location.long.x, 
                                                       lat1 = location.lat.x, 
                                                       lon2 = location.long.y, 
                                                       lat2 = location.lat.y, km = TRUE),
                                 var2 = individual.local.identifier)) 

havdf <- bind_rows(hav_km) %>%
  distinct()
colnames(havdf) <-c ("gcd","individual.local.identifier.x" )


alldat <- left_join(ddur, havdf, by = c("individual.local.identifier.x"))


alldat <- alldat %>%
  mutate(gcd_class = case_when(
    gcd < 1000 ~ 1,
    gcd >1001 & gcd <3000 ~2, 
   gcd >3001 & gcd <6000 ~3, 
    gcd >6001 ~4,
    .default = NA
 ))


ggplot(alldat, aes(x = dur_days, y = gcd, colour=  as.factor(dur_class)))+ 
  geom_point()
  

```

The great circle distance was calculated using the rough first and last location points as a rough proxy for the total movement. 

```{r, echo = FALSE}

gcddur_plot <- ggplot(alldat, aes(y=factor(tag.local.identifier))) +
  geom_segment(aes(x=0, xend=gcd, y=factor(tag.local.identifier), yend=factor(tag.local.identifier)), size=1)+
  xlab("Distance") + ylab("Tag") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10))

gcddur_plot

# gcp_hist <- ggplot(alldat, aes(x= gcd, colour = as.factor(year)))+
#   geom_histogram() + #fill="white", position="dodge") +
#   #facet_wrap(~year)+
#   xlab("distance (km)") 
# gcp_hist

gcp_hist <- ggplot(alldat, aes(x= gcd))+
  geom_histogram(bins = 50) + #fill="white", position="dodge") +
  #facet_wrap(~year)+
  xlab("distance (km)") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10))
gcp_hist

```

# Geographic distributon of tags

```{r, echo=FALSE}
  
rf_sf <- sf::st_as_sf(clean, coords = c("location.long","location.lat"), crs = 4326, agr = "constant")

world <- ne_countries(scale = "medium", returnclass = "sf")

Americas <- world %>% dplyr::filter(region_un == "Americas")
#Americas <- world %>% dplyr::filter(continent == "North America")

# entire north America 
global <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = rf_sf, size = 1, colour = "dark blue") +
  #facet_wrap(~tag.local.identifier)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

global

# 
# jpeg(file.path(out.plots,"rose_density.jpg"), width = 30, height = 30,units = "cm", res = 210)
# # 2. Create the plot
# global 
```
# Plot by duration class 

```{r, echo = FALSE}

#head(alldat)

all <- alldat %>% dplyr:: select(individual.local.identifier.x,dur_days, year, active, dur_class,gcd, gcd_class)%>%
  rename("individual.local.identifier"  = individual.local.identifier.x )
   
arf_sf <- left_join(rf_sf, all) 

                        
# entire north America 
dd <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = arf_sf, size = 1, colour = "dark blue") +
  facet_wrap(~dur_class)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

#dd


# entire north America 
gg <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = arf_sf, size = 1, colour = "dark blue") +
  facet_wrap(~gcd_class)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

#gg


```

# Plot based on duration class

 
```{r, echo = FALSE}
dd
```

Plot based on great circle distance class

gcd < 1000 ~ 1,
gcd >1001 & gcd <3000 ~2, 
gcd >3001 & gcd <6000 ~3, 
gcd >6001 ~4

```{r}
gg
```


# Detailed geographic locations vs sub populations

We can estimate the subpopulation level for those individuals with longer duration of tag timing?
Using the estimated ranges
- South East US
- Tierra Del Fuego
- Northern South America
- Western Gulf and western Pacific. 

As many of these individuals were banded at Delaware bay, this limits the subpops to SE_US, TDF, N_SA, while the WGWP group are less likely to be detected this far east. 


```{r, echo = FALSE}

min_lat <- arf_sf %>% 
  cbind(st_coordinates(arf_sf))%>%
  st_drop_geometry() %>%
  dplyr::select(individual.local.identifier, Y)%>%
  group_by(individual.local.identifier)%>%
  slice_min(Y) %>%
  mutate(pop_class = case_when(
     Y < -10 ~ "TDF",
     Y > -10.0001 & Y < 30 ~ "N_SA",
     #Y >  20.01 & Y < 30 ~ "SE_USA",
     Y > 30.1 ~ "USA",
      .default = NA
))


arf <- arf_sf %>%
  left_join(min_lat, by = "individual.local.identifier" )


# entire north America 
subpop <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = arf, size = 1, colour = "dark blue") +
  facet_wrap(~pop_class)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

subpop

```

Due to the short duration of the tags it is difficult to assign any individuals which might be from the Southern USA or northern South America groups. There are a number of individuals who are from the TDF regions which reach the breeding grounds which could be compared to the more southerly groups. 


This figure is somewhat interesting as it shows that the TDF individuals are breeding on the western side of arctic, while the SA (NA) is breeding on the eastern edge of the arctic. The USE or unknown group appear to be breeding somewhere in the middle, unless there is something else to separate the individuals. 



The 2023 dataset is on-going so more data is likely to come with time. 

There seems to be some variation with the breeding location with various islands being used. 

```{r}

# list of tags reviewed in QGIS: 


a <- sort(unique(clean$tag.local.identifier ))

#prince_charles & baffin island
pc <- c(230299, 230309, 230312, 232981, 234236,234238 ,240161, 242570) 

# unusual/non-breed
odd <- c( 230302, 230307, 230300,230315, 230316 )
         
nbreed <- c( 230306,  230319, 232980, 234235, 238543, 240157,240160, 240162, 240163, 
            240166, 240169, 242571,  240155, 238542, 
            240165) 

#southhampton
sth <- c(230308, 238544)  
        
#king william 
kw <- c(230301, 230304, 230310 , 230311, 230314 ,230317 , 234237, 240156, 240164, 240168, 242573,  242580) 
  
#princess Victoria     
pvic <- c (230303, 230318, 230320, 232982, 232985, 232986, 234233,  234234, 234239,234240, 238546,
           240158, 240159 , 240167, 241166,  241167, 242574, 242577) 
           

# breeding local 

breedsf <- arf_sf %>% 
  mutate(breed_class = case_when(
     tag.local.identifier %in%  pc ~ "prince_baffin",
     tag.local.identifier %in%  odd ~ "hudson_bay",
     tag.local.identifier %in% nbreed ~ "non_breed",
     tag.local.identifier %in%  sth ~ "Southhampton",
     tag.local.identifier %in%  kw ~ "king_william",
     tag.local.identifier %in%  pvic ~ "prince_vic",
      .default = NA
))


# entire north America 
breed_loc <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = breedsf, size = 1, colour = "dark blue") +
  facet_wrap(~breed_class)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

breed_loc
     

```


There was also an individual who flew way out to sea and then returned : #238544

This individual was in the southhampton nesting group. 









Questions/Notes 

- still missing reference data for number of birds


```{r}

ids <- brep %>% 
  dplyr::select(tag.local.identifier, animal.id) %>%
  distinct()

# based on the reference data there is 
# 107 unique tag ids (numeric) and 77 unique animal ids *ie NBRI_233781_2022

# read in track data

btemp <- read.csv(file.path(raw_dat, filesoi))
bout <- btemp %>%
    dplyr::select(visible, timestamp, location.long, location.lat,
                  gps.fix.type.raw, lotek.crc.status,
                  sensor.type, individual.local.identifier,
                  argos.lc, algorithm.marked.outlier,
                  tag.local.identifier)


# calculate time differences
bout <- bout %>%
  mutate(date_time = ymd_hms(timestamp))%>%
  mutate(year = year(date_time )) %>%
  mutate(month = month(date_time ),
         day = day(date_time),
         hour = hour(date_time),
         minute = minute(date_time))


# which ids are missing tracking data? 

# based on the tracks data there are 57 ids and animal tag nos. 
# all tracks have a animal tag id (ie: )

ids.track <- bout %>% 
  dplyr::select(tag.local.identifier, individual.local.identifier) %>%
  distinct()

fullid <- full_join(ids, ids.track)



```


Appendices: 

List of tags ids with no tracks or reference tags 

```{r, echo = FALSE}

norefnotrack = fullid %>%
  filter(is.na(individual.local.identifier))%>%
  filter(animal.id == "") %>% 
  pull(tag.local.identifier)
  
norefnotrack
length(norefnotrack)

```

List of tags ids with no tracks

```{r, echo = FALSE}
refnotrack = fullid %>%
  filter(is.na(individual.local.identifier))%>%
  filter(animal.id != "") %>% 
  pull(tag.local.identifier)
  
refnotrack
length(refnotrack)
#20
