==> 01_Summary_SouthCarolina_Arctic.Rmd <==
---
title: "01_Summary_SouthCarolina_Arctic"
author: "Gen Perkins"
date: "2024-01-05"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

## Summary of South Carolina to Arctic 

```{r library, warning=FALSE, message=FALSE, include = FALSE}
library(leaflet)
library(RColorBrewer)
library(dplyr)
library(lubridate)
library(sf)
library(adehabitatHR)
library(ggplot2)
library(stringr)
library(readxl)
library(dplyr)
library("rnaturalearth")
library("rnaturalearthdata")


raw_dat <- file.path("data", "movebank_locations_20231219")

filesoi <- list.files(raw_dat)

# data_set3 : spring migration 
key = "South Carolina"

filesoi <- list.files(raw_dat, pattern = key)
filesoi_ref <- filesoi[1]
filesoi <- filesoi[2]

# read in reference data 
brep <- read.csv(file.path(raw_dat, filesoi_ref))
brep <- brep %>%
  dplyr::select(c(tag.id , "animal.id" , deploy.on.date, animal.life.stage, tag.model, animal.sex,
                  deployment.comments,tag.manufacturer.name )) %>% 
  rename("animal.ring.id" = animal.id,
         "animal.marker.id" = deployment.comments) %>%
  mutate(study.site = "Kiawah")



brep  <- brep [complete.cases(brep ), ]

brep <- brep %>%
mutate(tag.local.identifier = tag.id)

btemp <- read.csv(file.path(raw_dat, filesoi))
bout <- btemp %>%
    dplyr::select(visible, timestamp, location.long, location.lat,
                  #gps.fix.type.raw, lotek.crc.status,
                  sensor.type, individual.local.identifier,
                  argos.lc, argos.altitude, #algorithm.marked.outlier,
                  tag.local.identifier)

# calculate time differences
bout <- bout  %>% mutate(date_time = ymd_hms(timestamp)) %>%
  mutate(year = year(date_time )) %>%
  mutate(month = month(date_time ),
         day = day(date_time),
         hour = hour(date_time),
         minute = minute(date_time))

#no.tags <- unique(bout$tag.local.identifier)
#no.ids <- unique(bout$individual.local.identifier)
  
# 
# ids <- bout %>%
#   dplyr::select(tag.local.identifier, individual.local.identifier) %>%
#   group_by(tag.local.identifier, individual.local.identifier)%>%
#   summarise(n = n())
#   
# 
# ymin = min(range(bout$year))
# ymax = max(range(bout$year))


# merge these together and output 

all_dat <- left_join(bout, brep) %>%
  dplyr::mutate(argos.lc = as.character(argos.lc)) %>%
  dplyr::select(-individual.local.identifier, -tag.local.identifier)%>%
  mutate(tag.id = as.character(tag.id)) %>%
  filter(location.long >= -120,
         location.long <= -62) %>% 
  mutate(tag.model = "Sunbird Solar Argos") %>%
  filter(!is.na(location.long), 
         !is.na(location.lat)) %>%
  mutate(deploy.on.latitude = 32.53945, 
         deploy.on.longitude = -80.17069)%>%
  dplyr::filter(argos.lc != "Z")%>%
  dplyr::filter(argos.lc != "") 

all_dat <- all_dat %>%
  mutate(id = seq(1, length(all_dat$visible), 1))

#unique(all_dat$argos.lc)

#type_dur <-all_dat|> 
#  dplyr::select(tag.id, argos.lc) |> 
#  group_by(argos.lc, tag.id) |> 
#  summarise( n =  n())

##############################################################################
## Calculate the stopoverlocations: 

# Duration between pings/ 
bdd <- all_dat |> 
  mutate(ddate = ymd_hms(date_time)) |> 
  arrange(tag.id, ddate)

bdd_dur <- bdd  |> 
  group_by(tag.id) |> 
  mutate(diff = difftime(ddate, lag(ddate),  units = c("hours")), 
         diff = as.numeric(diff))%>%
  dplyr::filter(diff >0)

# 
# type_dur <- bdd_dur |> 
#   dplyr::select(tag.id, argos.lc, diff) |> 
#   group_by(argos.lc, tag.id) |> 
#   summarise( n =  n(), 
#              min = min(diff, na.rm= T),
#              max =max(diff, na.rm = T), 
#              mean = mean(diff, na.rm = T))

## Calculate distance between points and bearing

bdd_det <- bdd_dur  |> 
  #filter(tag.id == 230318) |> 
  group_by(tag.id) |> 
  mutate(location.long_prior = lag(location.long, 1L),
         location.lat_prior = lag(location.lat, 1L)) %>%
  rowwise() %>%
  mutate(gcd_m = distHaversine(c(location.long_prior,location.lat_prior), c(location.long, location.lat)),
         bearing = bearing(c(location.long_prior,location.lat_prior), c(location.long, location.lat)),
         speed_mhr = round((gcd_m/diff)/1000,1))%>%
  ungroup()

# determine the location of direction 

head(bdd_det)


# stop over = within 25 km of previous point? 

bt <- bdd_det |> 
  dplyr::select(id, tag.id, ddate, day, month, year,location.long, location.lat,location.lat_prior, diff, gcd_m,bearing, speed_mhr) %>%
  group_by(tag.id)%>%
  mutate(stopover = ifelse( gcd_m <= 25000, "stop-over", "migration")) %>%
  mutate(breeding = case_when(
    stopover == "stop-over" & month %in% c(6,7,8) & location.lat > 60 ~ "breeding", 
    .default = "NA")) %>%
 mutate(direction = case_when(
    location.lat >= location.lat_prior ~ "northward", 
    location.lat <= location.lat_prior~ "southward",
    .default = "NA"))%>%
  ungroup()

clean_sf <- st_as_sf(bt, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path("output", "pt_sth_20240123.gpkg"), append = F)

# 
# 
# activity = case_when(
#     
#   )
#            )
# 

#########################################################\
# tag 1 
#########################################################


bt242656 <- bt %>% 
  dplyr::filter(tag.id == "242656")

idll <- bt242656 |> 
  dplyr::select("location.long", "location.lat", id)

clean_sf <- st_as_sf(bt242656, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path("output", "pt_bt242656.gpkg"), append = F)

# read back in and fix corrections; 

manual_edits <- file.path("output", "manual_edited_complete")
bt24edit <- read.csv(file.path( manual_edits, "pt_sth232565all.csv")) %>%
  mutate(breeding = case_when(
           stopover == "stop-over" & month %in% c(9, 10, 11, 12) ~ "wintering", 
           stopover == "stop-over" & month == 5 ~ "wintering", 
          .default = as.character(breeding)))%>%
  mutate(direction = case_when(
           breeding == "wintering"  ~ NA, 
           stopover == "breeding" ~ NA,
          .default = as.character(direction)))%>% 
  left_join(idll) %>%
  dplyr::filter(!is.na(location.lat))
  


clean_sf <- st_as_sf(bt24edit, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path(manual_edits, "pt_bt242656_final.gpkg"), append = F)


 #list.files(file.path("output", "manual_edited_complete"))



#clean_sf <- st_as_sf(bdd_det, coords = c("location.long", "location.lat"), crs = st_crs(4326))
#st_write(clean_sf, file.path("output", "pt_sth_20240123.gpkg"), append = F)




#########################################################\
# tag 2
#########################################################

bt242657 <- bt %>% 
  dplyr::filter(tag.id == "242657")

idll <- bt242657 |> 
  dplyr::select("location.long", "location.lat", id)

clean_sf <- st_as_sf(bt242657, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path("output", "pt_bt242657.gpkg"), append = F)

# read back in and fix corrections; 

manual_edits <- file.path("output", "manual_edited_complete")
bt67edit <- read.csv(file.path( manual_edits, "pt_sth232567all.csv")) %>%
  left_join(idll) %>%
  mutate(breeding = case_when(
           stopover == "stop-over" & month %in% c(9, 10, 11, 12) ~ "wintering", 
           stopover == "stop-over" & month == 5 & location.lat <35 ~ "wintering", 
          .default = as.character(breeding)))%>%
  mutate(direction = case_when(
           breeding == "wintering"  ~ NA, 
           stopover == "breeding" ~ NA,
          .default = as.character(direction)))%>% 
  dplyr::filter(!is.na(location.lat))
  


clean_sf <- st_as_sf(bt67edit, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path(manual_edits, "pt_bt242657_final.gpkg"), append = F)


#########################################################\
# tag 3
#########################################################

bt242658 <- bt %>% 
  dplyr::filter(tag.id == "242658")

idll <- bt242658 |> 
  dplyr::select("location.long", "location.lat", id)

clean_sf <- st_as_sf(bt242658, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path("output", "pt_bt242658.gpkg"), append = F)



# read back in and fix corrections; 

manual_edits <- file.path("output", "manual_edited_complete")
bt68edit <- read.csv(file.path( manual_edits, "pt_sth242568all.csv")) %>%
  left_join(idll) %>%
  #mutate(breeding = case_when(
  #         stopover == "stop-over" & month %in% c(9, 10, 11, 12) ~ "wintering", 
  #         stopover == "stop-over" & month == 5 & location.lat <35 ~ "wintering", 
  #        .default = as.character(breeding)))%>%
  mutate(direction = case_when(
           breeding == "wintering"  ~ NA, 
           stopover == "breeding" ~ NA,
          .default = as.character(direction)))%>% 
  dplyr::filter(!is.na(location.lat))
  


clean_sf <- st_as_sf(bt68edit, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path(manual_edits, "pt_bt242658_final.gpkg"), append = F)





# merge together 


sth_out <- bind_rows(bt24edit, bt67edit, bt68edit)%>%
 dplyr::select(id, diff, bearing, speed_mhr,  stopover,  breeding, direction, gcd_m)

length(bt$id)%>%
  mutate(tag.id = as.character(tag.id))
length(sth_out$id)

btout <- left_join(all_dat, sth_out, by = "id")




# #save out file
clean_save = btout %>% mutate(proj = "sthcarolina_arctic")
saveRDS(clean_save, file = file.path("output", "rekn_sthcarolina_20240123.rds"))




clean_sf <- st_as_sf(clean_save, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path(manual_edits, "rekn_sthcarolina_20240113.gpkg"), append = F)






```

This dataset has `r length(no.tags)` tags (tag.local.identifier) and `r length(no.ids)` (individual.local.identifier). 

```{r, echo = FALSE}
knitr::kable(ids)

```

```{r, echo=FALSE}

clean <- bout %>%
  filter(visible == "true")%>%
  filter(!is.na(location.long)) %>%
  filter(!is.na(location.lat))

# #save out file
clean_save = clean %>% mutate(proj = "sthcarolina_arctic")
saveRDS(clean_save, file = file.path("output", "rekn_sthcarolina_20231219.rds"))



cleanargos <- clean %>% 
  dplyr::filter(argos.lc != "Z")%>%
  dplyr::filter(argos.lc != "A")%>% 
 dplyr::filter(argos.lc != "B")

#length(cleanargos$visible)

clean = cleanargos

errrors_tble <- clean %>%
  group_by(tag.local.identifier, year, visible) %>%
  summarise(count = n())


# #save out file
clean_save = clean %>% mutate(proj = "sthcarolina_arctic")
saveRDS(clean_save, file = file.path("output", "rekn_sthcarolina_20231102.rds"))



```


# Raw data

The number of raw records as of September 2023 is `r length(bout$tag.local.identifier)`.  The raw date range varies from 
`r min(range(bout$year))` to `r max(range(bout$year))`.


If we remove all the algorithm marked outliers (lotek = error, argos accuracy value = A, B, Z). We reduce the number of observations in the data set from `r length(bout$tag.local.identifier)` to `r length(clean$tag.local.identifier)`.
Argos accuracy codes are: 

- Class 0: over 1500 m radius (not currently filtered)

- Class A: Argos no accuracy estimation (3 messages received)
- Class B: Argos no accuracy estimation (1 or 2 messages received)
- Class Z: Argos Invalid Location


The cleaned date range varies from 
`r min(range(clean$year))` to `r max(range(clean$year))`.


# GPS vs argos


The data split between gps and argos data records. Large proportion are argos doppler shift.   These represent two different methods of calculating the locations and all types can be used. 

```{r, echo = FALSE, message=FALSE}
tag_type <- clean %>%
  dplyr::select(tag.local.identifier, sensor.type) %>%
  group_by(tag.local.identifier, sensor.type)%>%
  summarise(n = n())

#tag_type

tag_type_date <- clean %>%
  dplyr::select(tag.local.identifier, sensor.type, year, month) %>%
   group_by(tag.local.identifier, sensor.type, year,month)%>%
  summarise(n = n())

#tag_type_date

#p_alldat <- ggplot(clean, aes(year, fill = sensor.type))+
#  geom_bar(position = "dodge") 

p_alldat <- ggplot(clean, aes(year, fill = sensor.type))+
  geom_bar(position = "dodge") +
  xlim(2021, 2024)+
  facet_wrap(~tag.local.identifier)+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

```
