==> 01_Summary_spring_USFWS.rmd <==
---
title: "01_Summary_Spring Migration/USFWS"
author: "Gen Perkins"
date: "2023-06-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Summary of Spring Migration/USFWS

```{r library, warning=FALSE, message=FALSE, include = FALSE}
library(leaflet)
library(RColorBrewer)
library(dplyr)
library(lubridate)
library(sf)
library(adehabitatHR)
library(ggplot2)
library(stringr)
library(readxl)
library(dplyr)
library("rnaturalearth")
library("rnaturalearthdata")


#raw_dat <- file.path("data", "movebank_locations_20230407")
raw_dat <- file.path("data", "movebank_locations_20231219")

filesoi <- list.files(raw_dat)

# data_set3 : spring migration 
key = "Spring Migration"

filesoi <- list.files(raw_dat, pattern = key)

filesoi_ref <- filesoi[1]
filesoi <- filesoi[2]


# read in reference data 
brep <- read.csv(file.path(raw_dat, filesoi_ref))
brep <- brep %>%
  mutate(tag.local.identifier = tag.id)

# number of unique tags = no of ids 

btemp <- read.csv(file.path(raw_dat, filesoi))
bout <- btemp %>%
    dplyr::select(visible, timestamp, location.long, location.lat,
                  argos.altitude, import.marked.outlier, 
                  gps.fix.type.raw, lotek.crc.status, 
                  sensor.type, individual.local.identifier,
                  argos.lc, algorithm.marked.outlier,
                  tag.local.identifier, height.above.ellipsoid)%>%
  mutate(date_time = ymd_hms(timestamp))%>%
  mutate(year = year(date_time )) %>%
  mutate(month = month(date_time ),
         day = day(date_time),
         hour = hour(date_time),
         minute = minute(date_time))

# no.tags <- unique(bout$tag.local.identifier)
# no.ids <- unique(bout$individual.local.identifier)
#   
# ids <- bout %>%
#   dplyr::select(tag.local.identifier, individual.local.identifier) %>%
#   group_by(tag.local.identifier, individual.local.identifier)%>%
#   summarise(n = n())
#   
# ymin = min(range(bout$year))
# ymax = max(range(bout$year))
#
# brep #107
# bout

all_dat <- left_join(bout, brep ) %>%
   dplyr::mutate(lotek.crc.status = case_when(
    is.na(lotek.crc.status) ~ "",
    TRUE ~ as.character(lotek.crc.status)))%>%
  dplyr::filter(lotek.crc.status != "E")%>%
  dplyr::mutate(argos.lc = case_when(
    is.na(argos.lc) ~ "",
    TRUE ~ as.character(argos.lc)))%>%
  dplyr::filter(argos.lc != "Z") %>%
  dplyr::mutate(deploy.on.latitude = case_when(
    study.site == "THOMPS" ~ 39.20075,
    study.site == "NORBURYC" ~ 39.20075)) %>%
  dplyr::mutate(deploy.on.longitude = case_when(
    study.site == "THOMPS" ~ -75.0255,
    study.site == "NORBURYC" ~ -75.0255)) %>%
  dplyr::select(-tag.local.identifier, -individual.local.identifier, -height.above.ellipsoid) %>%
  dplyr::mutate(argos.lc = as.character(argos.lc))%>%
  mutate(tag.id = as.character(tag.id)) %>%
   mutate(tag.model = "Lotek PinPoint GPS-Argos 75",
          tag.manufacturer.name = "Lotek")|> 
  filter(location.lat < 90 ) |> 
  filter(location.lat> -90)


head(all_dat)

all_dat <- all_dat %>%
  mutate(id = seq(1, length(all_dat$visible), 1))
# 
# 
# # Duration between pings/ 
# bdd <- all_dat |> 
#   mutate(ddate = ymd_hms(date_time)) |> 
#   arrange(tag.id, ddate)
# 
# bdd_dur <- bdd  |> 
#   group_by(tag.id) |> 
#   mutate(diff = difftime(ddate, lag(ddate),  units = c("hours")), 
#          diff = as.numeric(diff))%>%
#   dplyr::filter(diff >0)
# 
# 
# ## Calculate distance between points and bearing
# 
# bdd_det <- bdd_dur  |> 
#   #filter(tag.id == 230318) |> 
#   group_by(tag.id) |> 
#   mutate(location.long_prior = lag(location.long, 1L),
#          location.lat_prior = lag(location.lat, 1L)) %>%
#   rowwise() %>%
#   mutate(gcd_m = distHaversine(c(location.long_prior,location.lat_prior), c(location.long, location.lat)),
#          bearing = bearing(c(location.long_prior,location.lat_prior), c(location.long, location.lat)),
#          speed_mhr = round((gcd_m/diff)/1000,1))%>%
#   ungroup()
# 
# # determine the location of direction 
# 
# head(bdd_det)
# # 
# 
# # stop over = within 25 km of previous point? 
# 
# bt <- bdd_det |> 
#   dplyr::select(id, tag.id, ddate, day, month, year,location.long, location.lat,location.lat_prior, diff, gcd_m,bearing, speed_mhr) %>%
#   group_by(tag.id)%>%
#   mutate(stopover = ifelse( gcd_m <= 25000, "stop-over", "migration")) %>%
#   mutate(breeding = case_when(
#     stopover == "stop-over" & month %in% c(6,7,8) & location.lat > 60 ~ "breeding", 
#     .default = "NA")) %>%
#  mutate(direction = case_when(
#     location.lat >= location.lat_prior ~ "northward", 
#     location.lat <= location.lat_prior~ "southward",
#     .default = "NA"))%>%
#   ungroup()
# 
# 
# idll <- bt|> 
#   dplyr::select("location.long", "location.lat", id)
# 
# 
# clean_sf <- st_as_sf(bt, coords = c("location.long", "location.lat"), crs = st_crs(4326))
# st_write(clean_sf, file.path("output", "pt_spring_20240123.gpkg"), append = F)
# 
# # 
# 
# ## Post manual edits
# 
# manual_edits <- file.path("output", "manual_edited_complete")
# milsedit <- read.csv(file.path( manual_edits, "pt_mils_mig_all.csv")) %>%
#   left_join(idll) %>%
#    mutate(direction = case_when(
#            breeding == "wintering"  ~ NA, 
#            stopover == "breeding" ~ NA,
#           .default = as.character(direction)))%>% 
#   dplyr::filter(!is.na(location.lat)) %>%
#   dplyr::select(-tag.id, -ddate, -day, -month, -year, -"location.long", -"location.lat", -fid)
#  
# 
# head(milsedit)
# milout <- left_join(all_dat, milsedit, by = "id")
# 
# # #save out file
# clean_save = milout %>% mutate(proj = "Mispillion")
# saveRDS(clean_save, file = file.path("output", "rekn_mispillion_20240123.rds"))
# 
# 
# clean_sf <- st_as_sf(clean_save, coords = c("location.long", "location.lat"), crs = st_crs(4326))
# st_write(clean_sf, file.path(manual_edits, "rekn_mils_mig_20240123.gpkg"), append = F)
# 
# 
# 
# 






# #save out file
# #save out file
clean_save = all_dat %>% mutate(proj = "spring_USFWS")
saveRDS(clean_save, file = file.path("output", "rekn_spring_usfws_20240101.rds"))



```

This dataset has `r length(no.tags)` tags (tag.local.identifier) and `r length(no.ids)` (individual.local.identifier). 

For the reference data there are`r length(no.tags.ref)` tags and `r length(no.ids.ref)` (individual.local.identifier). 

```{r, echo = FALSE}
knitr::kable(ids)

```

```{r, echo=FALSE}

errors <- ggplot(bout, aes(year, fill = visible))+
    geom_bar(position = "dodge")#+
    #xlim(2021,2024)#+
  #facet_wrap(~tag.local.identifier)

clean <- bout %>%
  filter(visible == "true")%>%
  filter(!is.na(location.long)) %>%
  filter(!is.na(location.lat))  %>%
  filter(year <2024)


length(clean$visible)

#cleanargos <- clean %>% 
#  dplyr::filter(argos.lc != "Z")%>%
#  dplyr::filter(argos.lc != "A")%>% 
#  dplyr::filter(argos.lc != "B")

#length(cleanargos$visible)

#clean = cleanargos


```


# Raw data

The number of records as of early June 2023 is `r length(bout$tag.local.identifier)`.  The raw date range varies from 
`r min(range(bout$year))` to `r max(range(bout$year))`. Some obvious date errors to correct.


There seems to be lots of errors in this data set with accuracy. 




```{r}
errors

```





If we remove all the algorithm marked outliers (lotek = error, argos accuracy value = 0, A, B, Z). We reduce the number of observations in the data set from `r length(bout$tag.local.identifier)` to `r length(clean$tag.local.identifier)`.

The cleaned date range varies from 
`r min(range(clean$year))` to `r max(range(clean$year))`.



# GPS vs argos

The data split between gps and argos data records. 

```{r, echo = FALSE, message=FALSE}
tag_type <- clean %>%
  dplyr::select(tag.local.identifier, sensor.type) %>%
  group_by(tag.local.identifier, sensor.type)%>%
  summarise(n = n())

#tag_type

tag_type_date <- clean %>%
  dplyr::select(tag.local.identifier, sensor.type, year, month) %>%
   group_by(tag.local.identifier, sensor.type, year,month)%>%
  summarise(n = n())

#tag_type_date


p_alldat <- ggplot(clean, aes(year, fill = sensor.type))+
  geom_bar(position = "dodge") 

```

This seems odd as raw data shows shift from gps Lotek tag to argos-dopler-shift? 

```{r}
p_alldat 
```


# Duration of Tag summary

```{r, echo = FALSE}
table_max <- clean %>% 
  dplyr::filter(year < 2024) %>%
  dplyr::select(tag.local.identifier, date_time) %>%
 # group_by(tag.local.identifier)%>% 
  slice_max(date_time, by = tag.local.identifier)
colnames(table_max)<- c("tag.local.identifier", "max")

table_min <- clean %>% 
    dplyr::filter(year > 2020) %>%
  dplyr::select(tag.local.identifier, date_time) %>%
 # group_by(tag.local.identifier)%>% 
  slice_min(date_time, by = tag.local.identifier)

colnames(table_min)<- c("tag.local.identifier", "min")

dur <- left_join(table_max, table_min,by = join_by(tag.local.identifier)) %>%
  distinct()

#library(ggplot2)

dur_plot <- ggplot(dur, aes(y=factor(tag.local.identifier))) +
  geom_segment(aes(x=min, xend=max, y=factor(tag.local.identifier), yend=factor(tag.local.identifier)), size=2)+
  xlab("Date") + ylab("Tag") 

dur_plot

```




# Geographic distributon of tags

```{r, echo=FALSE}
rf_sf <- sf::st_as_sf(clean, coords = c("location.long","location.lat"), crs = 4326, agr = "constant")

world <- ne_countries(scale = "medium", returnclass = "sf")

Americas <- world %>% dplyr::filter(region_un == "Americas")
#Americas <- world %>% dplyr::filter(continent == "North America")
# entire north America 
global <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = rf_sf, size = 1, colour = "dark blue") +
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-110, -30), ylim = c(-10, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

global


# entire north America 
global <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = rf_sf, size = 1, colour = "dark blue") +
  facet_wrap(~tag.local.identifier)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-110, -30), ylim = c(-10, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

global

# jpeg(file.path(out.plots,"rose_density.jpg"), width = 30, height = 30,units = "cm", res = 210)
# # 2. Create the plot
# global 
```

Questions? 

- 15 tags with tracks and 16 tags in reference data set. This is due to low record number (ie only on record for 213840 and 213842), although see next comment. 
- These tags are duplicated tracks in the Atlantic Shores dataset although they do not seem to be an exact match (ie: more records in atlantic tracks dataset for the same tags?) This requires a bit more digging to figure out what is missing.
