==> 02_combine_all_data.Rmd <==
---
title: "02_Summary"
author: "Gen Perkins"
date: "2023-09-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

## Summary of all tracks: 


```{r}

#library(leaflet)
#library(RColorBrewer)

library(lubridate)
library(sf)
library(ggplot2)
library(stringr)
library(readxl)
library(dplyr)


raw_dat <- file.path("output")

filesoi <- list.files(raw_dat)

spring <- readRDS(file.path(raw_dat, "rekn_spring_usfws_20240129.rds")) #### WAITING ON MOVEBANK
at <- readRDS(file.path(raw_dat, "rekn_atlantic_20240129.rds")) #### WAITING ON MOVEBANK
eccc <- readRDS(file.path(raw_dat, "rekn_eccc_20240129.rds")) #### WAITING ON MOVEBANK

ma <- readRDS(file.path(raw_dat, "rekn_ma_mig_20240129.rds" ))        # edited for directions
dom <- readRDS(file.path(raw_dat, "rekn_dominion_20240129.rds"))      # edited for directions
oc <- readRDS(file.path(raw_dat, "rekn_ocean_20240129.rds"))          # edited for directions
mils <- readRDS(file.path(raw_dat, "rekn_mispillion_20240129.rds"))   # edited for directions
sth <- readRDS(file.path(raw_dat, "rekn_sthcarolina_20240129.rds"))   # edited for directions
qu <- readRDS(file.path(raw_dat, "rekn_mignon_raw_20240129.rds"))     # edited for directions
john <- readRDS(file.path(raw_dat, "rekn_john_20240129.rds"))         # edited for directions
new <- readRDS(file.path(raw_dat, "rekn_newstead_20240129.rds"))      # edited for directions

all <- bind_rows(ma, dom, oc, mils, sth, qu, john, new, eccc, at, spring) 

#all <- bind_rows(ma, john)
unique(all$proj)

unique(all$algorithm.marked.outlier)
unique(all$visible)
unique(all$import.marked.outlier)


# write out: 
write.csv(all, file.path("output", "xall_rekn_20240203.csv"), row.names = F)
clean_sf <- st_as_sf(all, coords = c("location.long", "location.lat"), crs = st_crs(4326))
st_write(clean_sf, file.path("output", "xall_rekn_20240203.gpkg"), append = F)


# extract reference information 
ref <- all %>% 
  dplyr::select(animal.comments, animal.id, animal.life.stage, animal.marker.id, 
                animal.mass,animal.ring.id, animal.sex, animal.taxon, "attachment.type", 
                "capture.timestamp",  "deploy.on.date" ,"deploy.on.latitude", "deploy.on.longitude" , 
                "deploy.on.measurements", "deploy.on.person" , "deployment.comment", 
                "duty.cycle", "sensor.type" , "tag.beacon.frequency"   ,   "tag.manufacturer.name",
                "tag.mass", "tag.id", "tag.serial.no" ,  "tag.model", "study.site", proj)   %>%
  distinct()
                                   
                
loc <- all %>% dplyr::select("proj", id, "tag.id", tag.model, "location.lat" ,"location.long" ,  
                         "date_time", "timestamp", # "year",   "month",    "day" ,   "hour" ,    "minute" , 
                        "argos.lc" , "lotek.crc.status", "gps.fix.type.raw" ,  
                        #"bearing" , "diff", "gcd_m",   "speed_mhr",  "breeding" ,"direction" ,
                        #"stopover", "location.lat_prior" , "location.long_prior",                     
                        "height.above.ellipsoid"  , "argos.altitude") %>%
  unique()
                                 
 # clean up the spatial resolution # can remove these cols. 
unique(loc$argos.lc)
unique(loc$lotek.crc.status)
unique(loc$gps.fix.type.raw)

## How many tags and types of tage 
unique(ref$tag.model)

#length(unique(ref$tag.id))
#29+66+1+5+22+43+42+10+15+15+2+36+34+3


### SUMMARY OF ANIMALS 

# tags per project 
tag.proj <-  ref %>% 
  dplyr::select(tag.id, proj) %>%
  group_by( proj) |> 
  summarise(no.of.tags = length(unique(tag.id)))

# mingnon tags not matching ##


# tag number per tag type
tag.types <- ref %>% 
  dplyr::select(tag.id, tag.model, proj) %>%
  group_by(tag.model, proj) |> 
  summarise(no.of.tags = length(unique(tag.id)))

# type of fix per tag model (accuracy per tag type - only relevant for pinpoint)
tag.types.model <- loc %>% 
  dplyr::select(tag.id, tag.model, proj, gps.fix.type.raw) %>%
  group_by(gps.fix.type.raw, tag.model) |> 
  summarise(no.of.tags = length(unique(tag.id)))
  

### Capture location 

study.site <- ref |> 
  dplyr::select(tag.id, study.site, deploy.on.date, deploy.on.latitude,  deploy.on.longitude ) |> 
  group_by(study.site) |> 
  summarise(no.of.tags = length(unique(tag.id)))

capture_loc <- ref |> 
  dplyr::select(tag.id, study.site,  animal.life.stage, animal.sex, deploy.on.date, deploy.on.latitude,  deploy.on.longitude ) %>%
  distinct() %>%
  st_as_sf( coords = c("deploy.on.longitude", "deploy.on.latitude"), crs = 4326)

# write out capture locations 
write_sf(capture_loc, file.path(raw_dat , "maps", "capture_locations.gpkg"))


## Age of individuals 

## capture age and sex 
sex_sum <- ref |> 
  dplyr::select(tag.id, study.site, proj, animal.life.stage, animal.sex) |> 
  group_by(animal.sex,proj) |> 
   summarise(no.of.tags = length(unique(tag.id)))

age_sum <- ref |> 
  dplyr::select(tag.id, study.site, proj, animal.life.stage) |> 
  group_by(animal.life.stage , study.site) |> 
  #group_by(animal.life.stage ) |> 
    summarise(no.of.tags = length(unique(tag.id)))



# ring-id for subpop information

rids <- ref %>%
  dplyr::select(animal.ring.id, tag.id, proj) %>% 
  group_by(animal.ring.id, tag.id ) %>%
  count()












# calculate the duration of the tags total 


dloc <- loc   %>%
  mutate(date_time = ymd_hms(timestamp))%>%
  mutate(year = year(date_time )) %>%
  mutate(month = month(date_time ),
         day = day(date_time),
         hour = hour(date_time),
         minute = minute(date_time)) %>%
  filter(year<=2023)

# # missing tag information 
# nas <- all %>%
#   filter(is.na(tag.model)) %>%
#   dplyr::select(proj) |>
#   distinct()
# 
# unique(all$tag.model)

















## Get list of Ring ids
# ring id 

rids <- dloc %>%
  dplyr::select(animal.ring.id, tag.id, proj) %>% 
  group_by(animal.ring.id)%>%
  count()



write.csv(rids, file.path("output", "rekn_ringid.csv"))


```


<!-- The cleaned date range varies from  -->
<!-- `r min(range(clean$year))` to `r max(range(clean$year))`. -->

<!-- ```{r} -->


<!-- raw_dat <- file.path("output") -->

<!-- bdat <- read.csv(file.path("output", "xsubset_rekn_20240119.csv"), row.names = F -->
<!-- #saveRDS(clean, file = file.path("output", "xall_rekn_20231102.rds")# The number of cleaned records we have compiled is `r length(clean$tag.local.identifier)`, from `r length(unique(clean$tag.local.identifier`) individuals. -->
<!-- #  -->
<!-- #  -->
<!-- # The cleaned date range varies from  -->
<!-- # `r min(range(clean$year))` to `r max(range(clean$year))`. -->



<!-- length(clean$tag.id) -->
<!-- length(unique(clean$tag.local.identifier)) -->


<!-- ``` -->

<!-- ```{r} -->
<!-- #  -->
<!-- # The number of cleaned records we have compiled is `r length(clean$tag.local.identifier)`, from `r length(unique(clean$tag.local.identifier`) individuals. -->
<!-- #  -->
<!-- #  -->
<!-- # The cleaned date range varies from  -->
<!-- # `r min(range(clean$year))` to `r max(range(clean$year))`. -->




<!-- table_max <- clean %>%  -->
<!--   dplyr::select(tag.local.identifier,individual.local.identifier, date_time) %>% -->
<!--  # group_by(tag.local.identifier)%>%  -->
<!--   slice_max(date_time, by = tag.local.identifier) -->
<!-- colnames(table_max)<- c("tag.local.identifier", "individual.local.identifier","max") -->

<!-- table_min <- clean %>%  -->
<!-- #   dplyr::filter(year != 2025) %>% -->
<!--   dplyr::select(tag.local.identifier,individual.local.identifier, date_time) %>% -->
<!--  # group_by(tag.local.identifier)%>%  -->
<!--   slice_min(date_time, by = tag.local.identifier) -->

<!-- colnames(table_min)<- c("tag.local.identifier", "individual.local.identifier", "min") -->

<!-- dur <- left_join(table_max, table_min, by = join_by(tag.local.identifier)) %>% -->
<!--   distinct() %>% -->
<!--   dplyr::mutate(duration = max- min)%>% -->
<!--   mutate(dur_min = round(as.numeric(duration)/60,1))%>% -->
<!--   mutate(dur_hrs = round(as.numeric(dur_min)/60,1))%>% -->
<!--   mutate(dur_days = round( dur_hrs/24,1))%>% -->
<!--   mutate(year = year(min)) -->

<!-- dur_data = dur -->
<!-- #library(ggplot2) -->

<!-- dur_plot <- ggplot(dur, aes(y=factor(tag.local.identifier))) + -->
<!--   geom_segment(aes(x=min, xend=max, y=factor(tag.local.identifier), yend=factor(tag.local.identifier)), size=1)+ -->
<!--   xlab("Date") + ylab("Tag")  -->

<!-- dur_plot -->

<!-- dur_hist <- ggplot(dur, aes(x= dur_days, colour = as.factor(year)))+ -->
<!--   geom_histogram() + #fill="white", position="dodge") + -->
<!--   facet_wrap(~year)+ -->
<!--   xlab("duration (days)")  -->

<!-- dur_hist -->

<!-- p_duration <- ggplot(clean, aes(factor(month), fill = factor(year)))+ -->
<!--   geom_bar(position = "dodge") + -->
<!--   #xlim(1,12)+ -->
<!--   #facet_wrap(~tag.local.identifier)+ -->
<!--   theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) -->

<!-- ``` -->

# great circle distance

```{r}

#remotes::install_github("mike-decarlo/gcd")
library(gcd)

table_last <- clean %>% 
  #dplyr::filter(year != 2025) %>%
  slice_max(date_time, by = tag.local.identifier) %>%
  slice_head(n = 1, by = tag.local.identifier) %>%
  dplyr::select( location.long, location.lat, individual.local.identifier)#%>%
 # mutate(timept = "end")

table_first <- clean %>% 
  #dplyr::filter(year != 2025) %>%
  slice_min(date_time, by = tag.local.identifier) %>%
  dplyr::select( location.long, location.lat, individual.local.identifier)#%>%
 # mutate(timept = "start")

loc_table = left_join(table_first, table_last, by = 'individual.local.identifier')
# 
# hav_km <- dist_haversine(lon1 = loc_table[1,1], lat1 = loc_table[1,2]
#   , lon2 = loc_table[1,4], lat2 = loc_table[1,5], km = TRUE)
# 
# hav_km
# 

library(tidyverse)
#df <- data.frame(a = 1:10, b = 11:20, c = 21:30)

hav_km <- loc_table %>% 
    group_nest(row_number()) %>% 
    pull(data) %>% 
    map(function(x) transmute(x,
                                 var1 = dist_haversine(lon1 = location.long.x, 
                                                       lat1 = location.lat.x, 
                                                       lon2 = location.long.y, 
                                                       lat2 = location.lat.y, km = TRUE),
                                 var2 = individual.local.identifier)) 

havdf <- bind_rows(hav_km) %>%
  distinct()
colnames(havdf) <-c ("gcd","individual.local.identifier.x" )


alldat <- left_join(dur, havdf, by = c("individual.local.identifier.x"))


alldat <- alldat %>%
  mutate(gcd_class = case_when(
    gcd < 1000 ~ 1,
    gcd >1001 & gcd <3000 ~2, 
   gcd >3001 & gcd <6000 ~3, 
    gcd >6001 ~4,
    .default = NA
 ))


ggplot(alldat, aes(x = dur_days, y = gcd))+ 
  geom_point()

```




# Geographic distributon of tags

```{r, echo=FALSE}
  
rf_sf <- sf::st_as_sf(clean, coords = c("location.long","location.lat"), crs = 4326, agr = "constant")
#st_write(rf_sf, file.path("output", "all_rekn_20230918.gpkg"))

world <- ne_countries(scale = "medium", returnclass = "sf")

Americas <- world %>% dplyr::filter(region_un == "Americas")
#Americas <- world %>% dplyr::filter(continent == "North America")

# entire north America 
global <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = rf_sf, size = 1, colour = "dark blue") +
  #facet_wrap(~tag.local.identifier)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

global


```


# Geographic distribution of all tracks with estimated groupings: 


- South_america n(22)
- Caribbean n(8)
- usa (70)
- west_arctic (37)
- east arctic (26)
- tdf (7)


These are estimateed groupings based on potential analysis questions. 


```{r}

a <- sort(unique(clean$tag.local.identifier ))

sth_am  <- c(204351, 204352, 204357, 204359,204364
             ,204369 ,204370, 204371, 204375, 224073, 224075,224083, 224085,
             224089, 224093, 224099, 233781, 233926  ,233928 ,236445,236450 ,236451)
 
#length(sth_am)
            
cari <- c(204361, 224080, 224082,224088 ,224097 ,233919,  236444, 236447)

#length(cari)

usa <- c(204362, 213839, 213842,221841 ,221843, 221844 ,221845, 221846 ,221847, 221850, 221852,
         221854 ,221856, 221858, 221860 ,221863, 221866, 221867 ,224072,
    224076 ,224077, 224078, 224079, 224081, 224086, 224087, 224091, 224092, 224094 ,224095, 224096, 224098, 224100, 224101, 224102,224103, 
    230306 ,230319, 232980 ,233918, 233920, 233921, 233922, 233923 ,233924, 233925,
    233929 ,233930, 233931 ,233932, 234177, 234178, 234179, 234180 ,234181, 234182, 234183 ,234184,
    234185 ,234187, 234189 ,234190, 234191,
     234235 ,236446, 236452, 238542, 238543, 240169, 242571)
     
#length(usa)

west_arctic <- c(213827, 213830,213834, 213836,213838, 213841, 230303, 230304,
   230310, 230311, 230314, 230317, 230318, 230320, 232982, 232985, 232986, 233927,
   234233, 234234, 234237,
   234239, 234240, 236453, 238546 ,240156, 240158, 240159 ,240167, 240168 ,240164, 241166 ,241167, 242573, 242574, 242577,230301)
   
#length(west_arctic)

east_arctic <- c(213828, 213829, 213831, 213832, 213833 ,213835, 213837, 213840, 230299, 230300, 230302,
230307,230308 ,230309, 230312 ,230315 ,230316,
232981,
 234236 , 234238 ,236448, 236449 ,238544 ,240161 ,242570 ,242580) 

#length(east_arctic)

tdf <- c(240155,  240157, 240160, 240162 ,240163, 240165 ,240166)
         
 #length(tdf) 

# breeding local 

breedsf <- rf_sf %>% 
  mutate(breed_class = case_when(
     tag.local.identifier %in%  tdf ~ "tdf",
     tag.local.identifier %in%  west_arctic ~ "west_arctic",
     tag.local.identifier %in% usa ~ "usa",
      tag.local.identifier %in%  sth_am ~ "South_america",
     tag.local.identifier %in%  east_arctic ~ "east_arctic",
      tag.local.identifier %in%  cari ~ "caribbean",
      .default = NA
))


# entire north America 
breed_loc <- ggplot(data = Americas) +
  geom_sf(color = "grey") +
  geom_sf(data = breedsf, size = 1, colour = "dark blue") +
  facet_wrap(~breed_class)+
  # geom_point(ru, aes(x = lng, y = lat), size = 4) +
  xlab("Longitude") + ylab("Latitude") +
  coord_sf(xlim = c(-130, -20), ylim = c(-50, 80), expand = FALSE)+
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank())

breed_loc
     

```
